Revert "rename the library (#18)"
This reverts commit ecfeed703a05a00f51c70a61ce3069a01f29a3c8.
