--- README.md
@@ -1,20 +1,0 @@
-# xarray-regridding
-
-**xarray-regridding** aims to regrid between different geospatial grids, while allowing to compute the grid weights in a distributed way.
-
-## Installation
-
-The library depends on:
-
-- xarray
-- numpy
-- sparse
-- opt_einsum
-- numba
-- scikit-learn
-
-Installation is currently only possible from github:
-
-```
-pip install git+https://github.com/iaocea/xarray-regridding.git
-```
