+++ ci/requirements/environment.yaml
@@ -1,4 +1,4 @@
-name: xarray-regridding-tests
+name: xarray-healpy-tests
 channels:
   - conda-forge
 dependencies:
