+++ pyproject.toml
@@ -1,5 +1,5 @@
 [project]
-name = "xarray-regridding"
+name = "xarray-healpy"
 requires-python = ">= 3.10"
 license = {text = "MIT"}
 dependencies = [
@@ -22,7 +22,7 @@ requires = ["setuptools>=64.0", "setuptools-scm"]
 build-backend = "setuptools.build_meta"
 
 [tool.setuptools]
-packages = ["xarray_regridding"]
+packages = ["xarray_healpy"]
 
 [tool.setuptools_scm]
 fallback_version = "9999"
@@ -63,7 +63,7 @@ fixable = ["I", "TID"]
 
 
 [tool.ruff.lint.isort]
-known-first-party = ["xarray_regridding"]
+known-first-party = ["xarray_healpy"]
 known-third-party = []
 
 [tool.ruff.lint.flake8-tidy-imports]
